requirements.txt
certifi==2018.1.18
chardet==3.0.4
click==6.7
diff-match-patch==20121119
Django==1.11
django-admin-honeypot==1.1.0
django-clear-cache==0.3
django-cors-headers==2.2.0
django-extensions==2.0.7
django-import-export==1.0.1
django-mongodb-engine==0.6.0
django-random-id-generator==0.1.0
django-template-debug==0.3.5
djangorestframework==3.7.7
djangotoolbox==1.8.0
djongo==1.1
et-xmlfile==1.0.1
Flask==0.12.2
gevent==1.2.2
graphene==2.0.1
graphene-django==2.0.0
graphql-core==2.0
graphql-relay==0.4.5
graphviz==0.8.3
greenlet==0.4.13
idna==2.6
iso8601==0.1.12
itsdangerous==0.24
jdcal==1.4
Jinja2==2.10
locustio==0.8.1
MarkupSafe==1.0
mongoengine==0.15.0
msgpack-python==0.5.6
odfpy==1.3.6
openpyxl==2.5.4
promise==2.1
psycopg2==2.7.4
pydotplus==2.0.2
pymongo==3.6.1
pyparsing==2.2.0
pytz==2018.3
PyYAML==3.12
pyzmq==17.0.0
requests==2.18.4
Rx==1.6.1
singledispatch==3.4.0.3
six==1.11.0
tablib==0.12.1
typing==3.6.4
unicodecsv==0.14.1
urllib3==1.22
Werkzeug==0.14.1
xlrd==1.1.0
xlwt==1.3.0
